backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
langchain==0.1.12
google-generativeai>=0.4.1,<0.5.0
langchain-google-genai==0.0.11 